requirements.txt
asgiref==3.11.1
coverage==7.13.3
Django==4.2.11
django-admin==2.0.2
django-cors-headers==4.9.0
django-excel-response2==3.0.6
django-filter==25.1
django-six==1.0.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
excel-base==1.0.4
gunicorn==25.0.3
isoweek==1.3.3
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.17.0
sqlparse==0.5.5
TimeConvert==3.0.13
tzlocal==5.3.1
xlwt==1.3.0
